+++ src/ai/cai_ethics_integration.cpp
@@ -111,7 +111,7 @@ std::vector<std::string> collectFormalizedPrinciples(
     const llm::ConstitutionalReasoningResult& cai_result) {
     std::vector<std::string> principles = {};
 
-    principles.reserve(cai_result.violated_principles.size() + static_cast<int>(cai_result.applied_principles.size()) );
+    principles.reserve(cai_result.violated_principles.size() + cai_result.applied_principles.size() );
 
     for (const auto& principle_id : cai_result.violated_principles) {
         appendUnique(principles, principle_id);
