+++ src/analytics/nlp_text_analyzer.cpp
@@ -287,7 +287,7 @@ std::vector<Keyword> NlpTextAnalyzer::extractKeywords(std::string_view text, siz
 
     // Sort by score and limit
     std::sort(keywords.begin(), keywords.end());
-    if (static_cast<int>(keywords.size()) > max_keywords) {
+    if (keywords.size() > max_keywords) {
         keywords.resize(max_keywords);
     }
 
@@ -397,7 +397,7 @@ ComplexityMetrics NlpTextAnalyzer::analyzeComplexity(std::string_view text) cons
 
     metrics.unique_words        = unique.size();
     metrics.avg_word_length     = static_cast<double>(total_length) / tokens.size();
-    metrics.avg_sentence_length = static_cast<double>(tokens.size()) / std::max<size_t>(1,static_cast<int>(sentences.size()));
+    metrics.avg_sentence_length = static_cast<double>(tokens.size()) / std::max<size_t>(1,sentences.size());
     metrics.lexical_diversity   = static_cast<double>(metrics.unique_words) / metrics.word_count;
 
     return metrics;
@@ -430,7 +430,7 @@ double NlpTextAnalyzer::estimateQueryComplexity(std::string_view query_text) con
 
     // Count tables
     auto tables = extractTableNames(query_text);
-    complexity += std::min(0.3,static_cast<int>(tables.size()) * 0.1);
+    complexity += std::min(0.3,tables.size() * 0.1);
 
     analysis_count_++;
     return std::min(1.0, complexity);
@@ -1700,15 +1700,15 @@ double NlpTextAnalyzer::calculateSimilarity(std::string_view text1, std::string_
         }
     }
 
-    size_t union_size = static_cast<int>(set1.size()) + static_cast<int>(set2.size()) - intersection;
+    size_t union_size = set1.size() + set2.size() - intersection;
     return union_size > 0 ? static_cast<double>(intersection) / union_size : 0.0;
 }
 
 std::map<std::string, size_t> NlpTextAnalyzer::getStatistics() const {
     return {{"analyses_performed", analysis_count_},
             {"tokens_processed", token_count_},
-            {"stopword_languages",static_cast<int>(stopwords_.size())},
-            {"sentiment_words",static_cast<int>(sentiment_lexicon_.size())}};
+            {"stopword_languages",stopwords_.size()},
+            {"sentiment_words",sentiment_lexicon_.size()}};
 }
 
 // ========== Private Helper Methods ==========
@@ -1900,7 +1900,7 @@ bool NlpTextAnalyzer::loadStopWordsFromYaml(const std::string &yaml_path, Langua
         }
 
         // Parse stop word (YAML list item: "  - word")
-        if (in_stopwords_section && static_cast<int>(line.size()) > 2 && line[0] == '-') {
+        if (in_stopwords_section && line.size() > 2 && line[0] == '-') {
             std::string word = line.substr(1);
             // Trim and remove quotes
             word.erase(0, word.find_first_not_of(" \t\"'"));
