+++ src/auth/mfa_authenticator.cpp
@@ -373,7 +373,7 @@ std::vector<uint8_t> MFAAuthenticator::hmacSHA1(
     
     HMAC(EVP_sha1(), 
          key.data(), static_cast<int>(key.size()),
-         message.data(), static_cast<int>(message.size()),
+         message.data(), message.size(),
          hash, &hash_len);
     
     return std::vector<uint8_t>(hash, hash + hash_len);
