+++ src/cache/bounded_lru_cache.cpp
@@ -311,7 +311,7 @@ bool BoundedLRUCache::contains(const std::string &key) const {
 
 std::size_t BoundedLRUCache::size() const {
     std::shared_lock<std::shared_mutex> lock(mutex_);
-    return static_cast<int>(cache_.size());
+    return cache_.size();
 }
 
 } // namespace cache
